+++ Digitalkirana/Views/Customer.cs
@@ -71,6 +71,11 @@ namespace Digitalkirana.Views
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
+            var result = MessageBox.Show("Do you really want to delete?", "Delete", MessageBoxButtons.OKCancel, MessageBoxIcon.Question);
+            if (result == DialogResult.Cancel)
+            {
+                return;
+            }
             customerDAL.DeleteCustomer(customer);
             reset();
             dataGridViewCustomer.DataSource = customerDAL.SelectAllCustomers();
