+++ Digitalkirana/Views/Product.cs
@@ -83,6 +83,11 @@ namespace Digitalkirana.Views
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
+            var result = MessageBox.Show("Do you really want to delete?", "Delete", MessageBoxButtons.OKCancel, MessageBoxIcon.Question);
+            if (result == DialogResult.Cancel)
+            {
+                return;
+            }
             productDAL.DeleteProduct(product);
             reset();
             dataGridViewProduct.DataSource = productDAL.SelectAllProducts();
@@ -124,6 +129,11 @@ namespace Digitalkirana.Views
 
         private void btnReset_Click(object sender, EventArgs e)
         {
+            var result = MessageBox.Show("Do you really want to reset?", "Reset", MessageBoxButtons.OKCancel, MessageBoxIcon.Question);
+            if (result == DialogResult.Cancel)
+            {
+                return;
+            }
             reset();
         }
